remove redundant code in _utils and move code from scatterplots to _utils. Rename color by the current number of colors
