+++ scanpy/plotting/_utils.py
@@ -262,60 +262,153 @@ def default_palette(palette=None):
     else: return palette
 
 
-def adjust_palette(palette: Union[Cycler, ListedColormap, Sequence[ColorLike]], length: int):
-    if isinstance(palette, cabc.Sequence):
-        length_pal = len(palette)
-    elif isinstance(palette, Cycler):
-        length_pal = len(palette.by_key()['color'])
-    elif isinstance(palette, ListedColormap):
-        length_pal = len(palette.colors)
-    else:
-        raise ValueError(f'Unknown palette type {type(palette)}')
-    if length_pal < length:
-        if length <= 28:
-            palette = palettes.default_26
-        elif length <= len(palettes.default_64):  # 103 colors
-            palette = palettes.default_64
-        else:
-            palette = ['grey' for i in range(length)]
-            logg.info("more than 103 colors would be required, initializing as 'grey'")
-    if isinstance(palette, (Cycler, ListedColormap, cabc.Sequence)):
-        return palette
-    else:
-        return cycler(color=palette)
+def _validate_palette(adata, key):
+    """
+    checks if the list of colors in adata.uns[f'{key}_colors'] is valid
+    and updates the color list in adata.uns[f'{key}_colors'] if needed.
+
+    Not only valid matplotlib colors are checked but also if the color name
+    is a valid R color name, in which case it will be translated to a valid name
+    """
+
+    _palette = []
+    color_key = f"{key}_colors"
+
+    for color in adata.uns[color_key]:
+        if not is_color_like(color):
+            # check if the color is a valid R color and translate it
+            # to a valid hex color value
+            if color in additional_colors:
+                color = additional_colors[color]
+            else:
+                logg.warning(
+                    f"The following color value found in adata.uns['{key}_colors'] "
+                    f"is not valid: '{color}'. Default colors will be used instead."
+                )
+                _set_default_colors_for_categorical_obs(adata, key)
+                _palette = None
+                break
+        _palette.append(color)
+    if _palette is not None:
+        adata.uns[color_key] = _palette
 
 
-def add_colors_for_categorical_sample_annotation(adata, key, palette=None, force_update_colors=False):
-    if key + '_colors' in adata.uns and not force_update_colors:
-        if len(adata.obs[key].cat.categories) > len(adata.uns[key + '_colors']):
-            logg.info(
-                f"    number of colors in `.uns[{key}'_colors']` smaller "
-                'than number of categories, falling back to palette'
-            )
-        else:
-            # make sure that these are valid colors
-            adata.uns[key + '_colors'] = [
-                additional_colors[c] if c in additional_colors else c
-                for c in adata.uns[key + '_colors']]
-            return
+def _set_colors_for_categorical_obs(adata, value_to_plot, palette):
+    """
+    Sets the adata.uns[value_to_plot + '_colors'] according to the given palette
+
+    Parameters
+    ----------
+    adata
+        annData object
+    value_to_plot
+        name of a valid categorical observation
+    palette
+        Palette should be either a valid :func:`~matplotlib.pyplot.colormaps` string,
+        a list of colors (in a format that can be understood by matplotlib,
+        eg. RGB, RGBS, hex, or a cycler object with key='color'
+
+    Returns
+    -------
+    None
+    """
+    from matplotlib.colors import to_hex
+    from cycler import Cycler, cycler
+
+    categories = adata.obs[value_to_plot].cat.categories
+    # check is palette is a valid matplotlib colormap
+    if isinstance(palette, str) and palette in pl.colormaps():
+        # this creates a palette from a colormap. E.g. 'Accent, Dark2, tab20'
+        cmap = pl.get_cmap(palette)
+        colors_list = [to_hex(x) for x in cmap(np.linspace(0, 1, len(categories)))]
+
     else:
-        logg.debug(f'generating colors for {key} using palette')
-    palette = default_palette(palette)
-    palette_adjusted = adjust_palette(palette,
-                                      length=len(adata.obs[key].cat.categories))
-    adata.uns[key + '_colors'] = palette_adjusted[
-        :len(adata.obs[key].cat.categories)].by_key()['color']
-    if len(adata.obs[key].cat.categories) > len(adata.uns[key + '_colors']):
-        raise ValueError(
-            'Cannot plot more than {} categories, which is not enough for {}.'
-            .format(len(adata.uns[key + '_colors']), key))
-    for iname, name in enumerate(adata.obs[key].cat.categories):
-        if name in settings.categories_to_ignore:
+        # check if palette is a list and convert it to a cycler, thus
+        # it doesnt matter if the list is shorter than the categories length:
+        if isinstance(palette, cabc.Sequence):
+            if len(palette) < len(categories):
+                logg.warning(
+                    "Length of palette colors is smaller than the number of "
+                    f"categories (palette length: {len(palette)}, "
+                    f"categories length: {len(categories)}. "
+                    "Some categories will have the same color."
+                )
+            # check that colors are valid
+            _color_list = []
+            for color in palette:
+                if not is_color_like(color):
+                    # check if the color is a valid R color and translate it
+                    # to a valid hex color value
+                    if color in _utils.additional_colors:
+                        color = _utils.additional_colors[color]
+                    else:
+                        raise ValueError("The following color value of the given palette is not valid: {}".format(color))
+                _color_list.append(color)
+
+            palette = cycler(color=_color_list)
+        if not isinstance(palette, Cycler):
+            raise ValueError("Please check that the value of 'palette' is a "
+                             "valid matplotlib colormap string (eg. Set2), a "
+                             "list of color names or a cycler with a 'color' key.")
+        if 'color' not in palette.keys:
+            raise ValueError("Please set the palette key 'color'.")
+
+        cc = palette()
+        colors_list = [to_hex(next(cc)['color']) for x in range(len(categories))]
+
+    adata.uns[value_to_plot + '_colors'] = colors_list
+
+
+def _set_default_colors_for_categorical_obs(adata, value_to_plot):
+    """
+    Sets the adata.uns[value_to_plot + '_colors'] using default color palettes
+
+    Parameters
+    ----------
+    adata : annData object
+    value_to_plot : name of a valid categorical observation
+
+    Returns
+    -------
+    None
+    """
+
+    categories = adata.obs[value_to_plot].cat.categories
+    length = len(categories)
+
+    # check if default matplotlib palette has enough colors
+    if len(rcParams['axes.prop_cycle'].by_key()['color']) >= length:
+        cc = rcParams['axes.prop_cycle']()
+        palette = [next(cc)['color'] for _ in range(length)]
+
+    else:
+        if length <= 20:
+            palette = palettes.default_20
+        elif length <= 28:
+            palette = palettes.default_28
+        elif length <= len(palettes.default_102):  # 103 colors
+            palette = palettes.default_102
+        else:
+            palette = ['grey' for _ in range(length)]
             logg.info(
-                f"    setting color of group {name!r} in {key!r} to 'grey' "
-                '(`sc.settings.categories_to_ignore`)'
+                f'the obs value {value_to_plot!r} has more than 103 categories. Uniform '
+                "'grey' color will be used for all categories."
             )
-            adata.uns[key + '_colors'][iname] = 'grey'
+
+    adata.uns[value_to_plot + '_colors'] = palette[:length]
+
+
+def add_colors_for_categorical_sample_annotation(adata, key, palette=None,
+                                                 force_update_colors=False):
+
+    color_key = f"{key}_colors"
+    colors_needed = len(adata.obs[key].cat.categories)
+    if palette and force_update_colors:
+        _set_colors_for_categorical_obs(adata, key, palette)
+    elif color_key in adata.uns and len(adata.uns[color_key]) <= colors_needed:
+        _validate_palette()
+    else:
+        _set_default_colors_for_categorical_obs(adata, key)
 
 
 def plot_edges(axs, adata, basis, edges_width, edges_color):
